+++ eqp/struct.jl
@@ -1,19 +1,19 @@
 #data structures used for equipment are specified in this file
 mutable struct elec
-   amp::Float64
-   volt::Float64
-   ohm::Float64
-   farrad::Float64
-   henry::Float64
-   yc::Float64
-   xl::Float64
+   amp::Float32
+   volt::Float32
+   ohm::Float32
+   farrad::Float32
+   henry::Float32
+   yc::Float32
+   xl::Float32
 end
 elec()=elec(69.69,69.69,69.69,69.69,69.69,69.69,69.69)
 
 #reliability structure
 mutable struct relia
-   fr::Float64
-   mttr::Float64
-   mc::Float64
+   fr::Float32
+   mttr::Float32
+   mc::Float32
 end
 relia()=relia(69.69,69.69,69.69)
